+++ backend/canisters/community/CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - When a member is removed also remove any invite ([#8328](https://github.com/open-chat-labs/open-chat/pull/8328))
+- On uninstall unsubscribe bot from community events ([#8330](https://github.com/open-chat-labs/open-chat/pull/8330))
 
 ## [[2.0.1829](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1829-community)] - 2025-07-09