+++ backend/canisters/community/CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Generalise Escrow so it can be used by 3rd parties ([#8356](https://github.com/open-chat-labs/open-chat/pull/8356))
 - Allow members to read community summary/membership ([#8357](https://github.com/open-chat-labs/open-chat/pull/8357))
 - `token:Cryptocurrency` -> `token_symbol:String` ([#8368](https://github.com/open-chat-labs/open-chat/pull/8368))
+- Use full names for bot notification fields ([#8389](https://github.com/open-chat-labs/open-chat/pull/8389))
 
 ### Fixed